+++ app/src/main/java/sg/edu/np/educaate1/Classes/Rating.java
@@ -4,9 +4,9 @@ public class Rating {
     private String ReviewId;
     private String userID;
     private String Desc;
-    private int Score;
+    private double Score;
 
-    public Rating(String id, String uid, String d, int s){
+    public Rating(String id, String uid, String d, double s){
         userID = uid;
         Desc = d;
         Score = s;
@@ -15,7 +15,7 @@ public class Rating {
 
     public  String getReviewId(){return ReviewId;}
 
-    public int getScore() {
+    public double getScore() {
         return Score;
     }
 
@@ -35,7 +35,7 @@ public class Rating {
 
     public void setUserID(String userid) { userID = userid; }
 
-    public void setScore(int score) {
+    public void setScore(double score) {
         Score = score;
     }
 }
